Project/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.7
django-cache-url==2.0.0
django-configurations==2.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-redis==4.8.0
djangorestframework==3.7.6
idna==2.6
kombu==4.1.0
Markdown==2.6.10
python-memcached==1.58
pytz==2017.3
redis==2.10.6
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
